requirements.txt
pandas>=2.0
numpy>=1.24
openpyxl>=3.1
scikit-learn>=1.3
xgboost>=2.0
matplotlib>=3.7
seaborn>=0.12
lifelines>=0.27
pyyaml>=6.0
